+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -585,6 +585,8 @@ export default class Meeting extends StatelessWebexPlugin {
   environment: string;
   namespace = MEETINGS;
   allowMediaInLobby: boolean;
+  localShareInstanceId: string;
+  remoteShareInstanceId: string;
   turnDiscoverySkippedReason: string;
   turnServerUsed: boolean;
   private retriedWithTurnServer: boolean;
@@ -1238,6 +1240,24 @@ export default class Meeting extends StatelessWebexPlugin {
      * @memberof Meeting
      */
     this.keepAliveTimerId = null;
+
+    /**
+     * id for tracking Local Share instances in Call Analyzer
+     * @instance
+     * @type {String}
+     * @private
+     * @memberof Meeting
+     */
+    this.localShareInstanceId = null;
+
+    /**
+     * id for tracking Remote Share instances in Call Analyzer
+     * @instance
+     * @type {String}
+     * @private
+     * @memberof Meeting
+     */
+    this.remoteShareInstanceId = null;
 
     /**
      * The class that helps to control recording functions: start, stop, pause, resume, etc
@@ -2508,6 +2528,8 @@ export default class Meeting extends StatelessWebexPlugin {
         switch (newShareStatus) {
           case SHARE_STATUS.REMOTE_SHARE_ACTIVE: {
             const sendStartedSharingRemote = () => {
+              this.remoteShareInstanceId = contentShare.shareInstanceId;
+
               Trigger.trigger(
                 this,
                 {
@@ -2518,7 +2540,7 @@ export default class Meeting extends StatelessWebexPlugin {
                 {
                   memberId: contentShare.beneficiaryId,
                   url: contentShare.url,
-                  shareInstanceId: contentShare.shareInstanceId,
+                  shareInstanceId: this.remoteShareInstanceId,
                   annotationInfo: contentShare.annotation,
                 }
               );
@@ -2555,6 +2577,7 @@ export default class Meeting extends StatelessWebexPlugin {
               name: 'client.share.floor-granted.local',
               payload: {
                 mediaType: 'share',
+                shareInstanceId: this.localShareInstanceId,
               },
               options: {meetingId: this.id},
             });
@@ -2597,6 +2620,8 @@ export default class Meeting extends StatelessWebexPlugin {
       } else if (newShareStatus === SHARE_STATUS.REMOTE_SHARE_ACTIVE) {
         // if we got here, then some remote participant has stolen
         // the presentation from another remote participant
+        this.remoteShareInstanceId = contentShare.shareInstanceId;
+
         Trigger.trigger(
           this,
           {
@@ -2607,7 +2632,7 @@ export default class Meeting extends StatelessWebexPlugin {
           {
             memberId: contentShare.beneficiaryId,
             url: contentShare.url,
-            shareInstanceId: contentShare.shareInstanceId,
+            shareInstanceId: this.remoteShareInstanceId,
             annotationInfo: contentShare.annotation,
           }
         );
@@ -5706,7 +5731,10 @@ export default class Meeting extends StatelessWebexPlugin {
       // @ts-ignore
       this.webex.internal.newMetrics.submitClientEvent({
         name: 'client.media.tx.start',
-        payload: {mediaType: data.type},
+        payload: {
+          mediaType: data.type,
+          shareInstanceId: data.type === 'share' ? this.localShareInstanceId : undefined,
+        },
         options: {
           meetingId: this.id,
         },
@@ -5716,7 +5744,10 @@ export default class Meeting extends StatelessWebexPlugin {
       // @ts-ignore
       this.webex.internal.newMetrics.submitClientEvent({
         name: 'client.media.tx.stop',
-        payload: {mediaType: data.type},
+        payload: {
+          mediaType: data.type,
+          shareInstanceId: data.type === 'share' ? this.localShareInstanceId : undefined,
+        },
         options: {
           meetingId: this.id,
         },
@@ -5735,7 +5766,10 @@ export default class Meeting extends StatelessWebexPlugin {
       // @ts-ignore
       this.webex.internal.newMetrics.submitClientEvent({
         name: 'client.media.rx.start',
-        payload: {mediaType: data.type},
+        payload: {
+          mediaType: data.type,
+          shareInstanceId: data.type === 'share' ? this.remoteShareInstanceId : undefined,
+        },
         options: {
           meetingId: this.id,
         },
@@ -5745,7 +5779,10 @@ export default class Meeting extends StatelessWebexPlugin {
       // @ts-ignore
       this.webex.internal.newMetrics.submitClientEvent({
         name: 'client.media.rx.stop',
-        payload: {mediaType: data.type},
+        payload: {
+          mediaType: data.type,
+          shareInstanceId: data.type === 'share' ? this.remoteShareInstanceId : undefined,
+        },
         options: {
           meetingId: this.id,
         },
@@ -6991,11 +7028,14 @@ export default class Meeting extends StatelessWebexPlugin {
       if (content && this.shareStatus !== SHARE_STATUS.LOCAL_SHARE_ACTIVE) {
         // @ts-ignore
         this.webex.internal.newMetrics.submitClientEvent({
-          name: 'client.share.initiated',
+          name: 'client.share.floor-grant.request',
           payload: {
             mediaType: 'share',
+            shareInstanceId: this.localShareInstanceId,
           },
-          options: {meetingId: this.id},
+          options: {
+            meetingId: this.id,
+          },
         });
 
         return this.meetingRequest
@@ -7024,6 +7064,19 @@ export default class Meeting extends StatelessWebexPlugin {
               locus_id: this.locusUrl.split('/').pop(),
               reason: error.message,
               stack: error.stack,
+            });
+
+            // @ts-ignore
+            this.webex.internal.newMetrics.submitClientEvent({
+              name: 'client.share.floor-granted.local',
+              payload: {
+                mediaType: 'share',
+                errors: MeetingUtil.getChangeMeetingFloorErrorPayload(error.message),
+                shareInstanceId: this.localShareInstanceId,
+              },
+              options: {
+                meetingId: this.id,
+              },
             });
 
             this.screenShareFloorState = ScreenShareFloorStatus.RELEASED;
@@ -7077,6 +7130,7 @@ export default class Meeting extends StatelessWebexPlugin {
         name: 'client.share.stopped',
         payload: {
           mediaType: 'share',
+          shareInstanceId: this.localShareInstanceId,
         },
         options: {meetingId: this.id},
       });
@@ -7957,6 +8011,17 @@ export default class Meeting extends StatelessWebexPlugin {
     }
 
     if (floorRequestNeeded) {
+      this.localShareInstanceId = uuid.v4();
+
+      // @ts-ignore
+      this.webex.internal.newMetrics.submitClientEvent({
+        name: 'client.share.initiated',
+        payload: {
+          mediaType: 'share',
+          shareInstanceId: this.localShareInstanceId,
+        },
+        options: {meetingId: this.id},
+      });
       // we're sending the http request to Locus to request the screen share floor
       // only after the SDP update, because that's how it's always been done for transcoded meetings
       // and also if sharing from the start, we need confluence to have been created